Commit 10
Almost ludum dare complete. Still need to move the origin in render to
body.

+++ EntityEngine/Components/Health.cs
@@ -10,7 +10,7 @@ namespace EntityEngine.Components
             HitPoints = hp;
         }
 
-        public int HitPoints { get; set; }
+        public float HitPoints { get; set; }
 
         public bool Alive
         {
@@ -21,7 +21,7 @@ namespace EntityEngine.Components
 
         public event Entity.EventHandler DiedEvent;
 
-        public void Hurt(int points)
+        public void Hurt(float points)
         {
             if (!Alive) return;
